# Conflicts:
#	bin/main/application.properties

+++ src/main/java/com/project/bookstore/web/Books/dto/BookListDto.java
@@ -19,7 +19,7 @@ public class BookListDto {
     private Long bookPri;
     private String bookGen;
     private String bookDet;
-    private int bookSto;
+    private Long bookSto;
 
     public BookListDto(Books entity) {
         this.isbn = entity.getIsbn();
